pyproject.toml
[project]
name = "fastapi-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-cloud-firestore>=2.21.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.7",
]
